Remove unused DrawContainer, extract CanvasHolder to separate file
